backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7.0
yagmail==0.15.293
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4.0